Trait rp2040_hal::clocks::Clock

pub trait Clock: Sealed + Sized {
    type Variant;

    fn freq(&self) -> HertzU32;
    fn configure_clock<S: ValidSrc<Self>>(
        &mut self,
        src: &S,
        freq: HertzU32
    ) -> Result<(), ClockError>;
}

For clocks

Required Associated Types

type Variant

Enum with valid source clocks register values for Clock

Required Methods

fn freq(&self) -> HertzU32

Get operating frequency

fn configure_clock<S: ValidSrc<Self>>(
&mut self,
src: &S,
freq: HertzU32
) -> Result<(), ClockError>

Configure this clock based on a clock source and desired frequency