Struct rp2040_hal::gpio::dynpin::DynPin

pub struct DynPin { /* private fields */ }

A value-level pin, parameterized by DynPinId and DynPinMode

This type acts as a type-erased version of Pin. Every pin is represented by the same type, and pins are tracked and distinguished at run-time.

Implementations

impl DynPin

pub fn id(&self) -> DynPinId

Return a copy of the pin ID

pub fn mode(&self) -> DynPinMode

Return a copy of the pin mode

pub fn try_into_mode(&mut self, mode: DynPinMode) -> Result<(), Error>

Convert the pin to the requested DynPinMode

pub fn into_floating_disabled(&mut self)

Disable the pin and set it to float

pub fn into_pull_down_disabled(&mut self)

Disable the pin and set it to pull down

pub fn into_pull_up_disabled(&mut self)

Disable the pin and set it to pull up

pub fn into_floating_input(&mut self)

Configure the pin to operate as a floating input

pub fn into_pull_down_input(&mut self)

Configure the pin to operate as a pulled down input

pub fn into_pull_up_input(&mut self)

Configure the pin to operate as a pulled up input

pub fn into_push_pull_output(&mut self)

Configure the pin to operate as a push-pull output

pub fn into_readable_output(&mut self)

Configure the pin to operate as a readable push pull output