Module rp2040_hal::gpio::pin

Type-level module for GPIO pins

Based heavily on atsamd-hal.

This module provides a type-level API for GPIO pins. It uses the type system to track the state of pins at compile-time. To do so, it uses traits to represent type classes and types as instances of those type classes. For example, the trait InputConfig acts as a type-level enum of the available input configurations, and the types Floating, PullDown, PullUp and BusKeep are its type-level variants.

When applied as a trait bound, a type-level enum restricts type parameters to the corresponding variants. All of the traits in this module are closed, using the Sealed trait pattern, so the type-level instances found in this module are the only possible variants.

Type-level Pins are parameterized by two type-level enums, PinId and PinMode.

A PinId identifies a pin by it’s group (BANK0 or QSPI) and pin number. Each PinId instance is named according to its datasheet identifier, e.g. Gpio0.

A PinMode represents the various pin modes. The available PinMode variants are Disabled, Input, Output and Function, each with its own corresponding configurations.

It is not possible for users to create new instances of a Pin. Singleton instances of each pin are made available to users through the Pins struct.

To create the Pins struct, users must supply the PAC IO_BANK0 and PAD_BANK0 peripherals as well as the SIO partition. The Pins struct takes ownership of the peripherals and provides the corresponding pins. Each Pin within the Pins struct can be moved out and used individually.

let mut peripherals = pac::Peripherals::take().unwrap();
let sio = Sio::new(peripherals.SIO);
let pins = Pins::new(peripherals.IO_BANK0,peripherals.PADS_BANK0,sio.gpio_bank0, &mut peripherals.RESETS);

Pins can be converted between modes using several different methods.

// Use one of the literal function names
let gpio12 = pins.gpio12.into_floating_input();
// Use a generic method and one of the `PinMode` variant types
let gpio12 = gpio12.into_mode::<FloatingInput>();
// Specify the target type and use `.into_mode()`
let gpio12: Pin<Gpio12, FloatingInput> = gpio12.into_mode();

Embedded HAL traits

This module implements all of the embedded HAL GPIO traits for each Pin in the corresponding PinModes, namely: InputPin, OutputPin, ToggleableOutputPin and StatefulOutputPin.

For example, you can control the logic level of an OutputPin like so

use rp2040_hal::{pac, gpio::{bank0::Gpio12, Pin, Pins, PushPullOutput}, sio::Sio};
use embedded_hal::digital::v2::OutputPin;

let mut peripherals = pac::Peripherals::take().unwrap();
let sio = Sio::new(peripherals.SIO);
let pins = Pins::new(peripherals.IO_BANK0,peripherals.PADS_BANK0,sio.gpio_bank0, &mut peripherals.RESETS);

let mut pin12: Pin<Gpio12, PushPullOutput> = pins.gpio12.into_mode();
pin12.set_high();

Type-level features

This module also provides additional, type-level tools to work with GPIO pins.

The OptionalPinId and OptionalPin traits use the OptionalKind pattern to act as type-level versions of Option for PinId and Pin respectively. And the AnyPin trait defines an AnyKind type class for all Pin types.

Re-exports

pub use bank0::Pins;

Modules

bank0

GPIO Pins for Bank0

qspi

GPIO Pins for Qspi

Structs

Disabled

Type-level variant of PinMode for disabled modes

Function

Type-level variant of PinMode for alternate peripheral functions

Input

Type-level variant of PinMode for input modes

Output

Type-level variant of PinMode for output modes

Pin

A type-level GPIO pin, parameterized by PinId and PinMode types

Enums

BusKeep

Type-level variant of both DisabledConfig and InputConfig

Clock

Type-level variant of FunctionConfig for alternate peripheral function Clock

Floating

Type-level variant of both DisabledConfig and InputConfig

I2C

Type-level variant of FunctionConfig for alternate peripheral function I2C

PinState

Digital output pin state

PullDown

Type-level variant of both DisabledConfig and InputConfig

PullUp

Type-level variant of both DisabledConfig and InputConfig

PushPull

Type-level variant of OutputConfig for a push-pull configuration

Pwm

Type-level variant of FunctionConfig for alternate peripheral function Pwm

Readable

Type-level variant of OutputConfig for a readable push-pull configuration

Spi

Type-level variant of FunctionConfig for alternate peripheral function Spi

Uart

Type-level variant of FunctionConfig for alternate peripheral function Uart

UsbAux

Type-level variant of FunctionConfig for alternate peripheral function UsbAux

Xip

Type-level variant of FunctionConfig for alternate peripheral function Xip

Traits

AnyPin

Type class for Pin types

DisabledConfig

Type-level enum for disabled configurations

FunctionConfig

Type-level enum for alternate peripheral function configurations

InputConfig

Type-level enum for input configurations

OptionalPin

Type-level equivalent of Option<PinId>

OptionalPinId

Type-level equivalent of Option<PinId>

OutputConfig

Type-level enum for output configurations

PinId

Type-level enum for pin IDs

PinMode

Type-level enum representing pin modes

SomePin

Type-level equivalent of Some(PinId)

SomePinId

Type-level equivalent of Some(PinId)

ValidPinMode

Type-level marker for tracking which pin modes are valid for which pins

Type Definitions

BusKeepDisabled

Type-level variant of PinMode for bus keep disabled mode

BusKeepInput

Type-level variant of PinMode for bus keep input mode

FloatingDisabled

Type-level variant of PinMode for floating disabled mode

FloatingInput

Type-level variant of PinMode for floating input mode

FunctionClock

Type-level variant of PinMode for alternate peripheral function Clock

FunctionI2C

Type-level variant of PinMode for alternate peripheral function I2C

FunctionPio0

Type-level variant of PinMode for alternate peripheral function pac::PIO0

FunctionPio1

Type-level variant of PinMode for alternate peripheral function pac::PIO1

FunctionPwm

Type-level variant of PinMode for alternate peripheral function Pwm

FunctionSpi

Type-level variant of PinMode for alternate peripheral function Spi

FunctionUart

Type-level variant of PinMode for alternate peripheral function Uart

FunctionUsbAux

Type-level variant of PinMode for alternate peripheral function UsbAux

FunctionXip

Type-level variant of PinMode for alternate peripheral function Xip

PullDownDisabled

Type-level variant of PinMode for pull-down disabled mode

PullDownInput

Type-level variant of PinMode for pull-down input mode

PullUpDisabled

Type-level variant of PinMode for pull-up disabled mode

PullUpInput

Type-level variant of PinMode for pull-up input mode

PushPullOutput

Type-level variant of PinMode for push-pull output mode

ReadableOutput

Type-level variant of PinMode for readable push-pull output mode

SpecificPin

Type alias to recover the specific Pin type from an implementation of AnyPin