Struct rp2040_hal::gpio::pin::Pin

pub struct Pin<I, M> where
    I: PinId,
    M: PinMode + ValidPinMode<I>, { /* private fields */ }

A type-level GPIO pin, parameterized by PinId and PinMode types

Implementations

impl<I, M> Pin<I, M> where
I: PinId,
M: PinMode + ValidPinMode<I>,

pub fn into_mode<N: PinMode + ValidPinMode<I>>(self) -> Pin<I, N>

Convert the pin to the requested PinMode

pub fn into_floating_disabled(self) -> Pin<I, FloatingDisabled>

Disable the pin and set it to float

pub fn into_pull_down_disabled(self) -> Pin<I, PullDownDisabled>

Disable the pin and set it to pull down

pub fn into_pull_up_disabled(self) -> Pin<I, PullUpDisabled>

Disable the pin and set it to pull up

pub fn into_floating_input(self) -> Pin<I, FloatingInput>

Configure the pin to operate as a floating input

pub fn into_pull_down_input(self) -> Pin<I, PullDownInput>

Configure the pin to operate as a pulled down input

pub fn into_pull_up_input(self) -> Pin<I, PullUpInput>

Configure the pin to operate as a pulled up input

pub fn into_bus_keep_input(self) -> Pin<I, BusKeepInput>

Configure the pin to operate as a bus keep input

pub fn into_push_pull_output(self) -> Pin<I, PushPullOutput>

Configure the pin to operate as a push-pull output.

If you want to specify the initial pin state, use Pin::into_push_pull_output_in_state.

pub fn into_push_pull_output_in_state(
self,
state: PinState
) -> Pin<I, PushPullOutput>

Configure the pin to operate as a push-pull output, specifying an initial state which is applied immediately.

pub fn into_readable_output(self) -> Pin<I, ReadableOutput>

Configure the pin to operate as a readable push pull output.

If you want to specify the initial pin state, use Pin::into_readable_output_in_state.

pub fn into_readable_output_in_state(
self,
state: PinState
) -> Pin<I, ReadableOutput>

Configure the pin to operate as a readable push pull output, specifying an initial state which is applied immediately.

pub fn get_drive_strength(&self) -> OutputDriveStrength

Read the current drive strength of the pin.

pub fn set_drive_strength(&mut self, strength: OutputDriveStrength)

Set the drive strength for the pin.

pub fn get_slew_rate(&self) -> OutputSlewRate

Get the slew rate for the pin.

pub fn set_slew_rate(&mut self, rate: OutputSlewRate)

Set the slew rate for the pin.

pub fn clear_interrupt(&mut self, interrupt: Interrupt)

Clear interrupt.

pub fn interrupt_status(&self, interrupt: Interrupt) -> bool

Interrupt status.

pub fn is_interrupt_enabled(&self, interrupt: Interrupt) -> bool

Is interrupt enabled.

pub fn set_interrupt_enabled(&self, interrupt: Interrupt, enabled: bool)

Enable or disable interrupt.

pub fn is_interrupt_forced(&self, interrupt: Interrupt) -> bool

Is interrupt forced.

pub fn set_interrupt_forced(&self, interrupt: Interrupt, forced: bool)

Force or release interrupt.

pub fn set_interrupt_override(&mut self, override_value: InterruptOverride)

Set the interrupt override.

pub fn set_input_override(&mut self, override_value: InputOverride)

Set the input override.

pub fn set_output_enable_override(
&mut self,
override_value: OutputEnableOverride
)

Set the output enable override.

pub fn set_output_override(&mut self, override_value: OutputOverride)

Set the output override.