Struct rp2040_hal::pwm::Slice

pub struct Slice<I, M> where
    I: SliceId,
    M: SliceMode + ValidSliceMode<I>, {
    pub channel_a: Channel<I, M, A>,
    pub channel_b: Channel<I, M, B>,
    /* private fields */
}

Pwm slice

Fields
channel_a: Channel<I, M, A>

Channel A (always output)

channel_b: Channel<I, M, B>

Channel B (input or output)

Implementations

impl<I, M> Slice<I, M> where
I: SliceId,
M: SliceMode + ValidSliceMode<I>,

pub fn into_mode<N: SliceMode + ValidSliceMode<I>>(self) -> Slice<I, N>

Convert the slice to the requested SliceMode

pub fn default_config(&mut self)

Set a default config for the slice

pub fn advance_phase(&mut self)

Advance the phase with one count

Counter must be running at less than full speed (div_int + div_frac / 16 > 1)

pub fn retard_phase(&mut self)

Retard the phase with one count

Counter must be running at less than full speed (div_int + div_frac / 16 > 1)

pub fn set_ph_correct(&mut self)

Enable phase correct mode

pub fn clr_ph_correct(&mut self)

Disables phase correct mode

pub fn enable(&mut self)

Enable slice

pub fn disable(&mut self)

Disable slice

pub fn set_div_int(&mut self, value: u8)

Sets the integer part of the clock divider

pub fn set_div_frac(&mut self, value: u8)

Sets the fractional part of the clock divider

pub fn get_counter(&self) -> u16

Get the counter register value

pub fn set_counter(&mut self, value: u16)

Set the counter register value

pub fn get_top(&self) -> u16

Get the top register value

pub fn set_top(&mut self, value: u16)

Sets the top register value

pub fn enable_interrupt(&mut self)

Enable the PWM_IRQ_WRAP interrupt when this slice overflows.

pub fn disable_interrupt(&mut self)

Disable the PWM_IRQ_WRAP interrupt for this slice.

pub fn has_overflown(&self) -> bool

Did this slice trigger an overflow interrupt?

pub fn clear_interrupt(&mut self)

Mark the interrupt handled for this slice.

pub fn force_interrupt(&mut self)

Force the interrupt. This bit is not cleared by hardware and must be manually cleared to stop the interrupt from continuing to be asserted.

pub fn clear_force_interrupt(&mut self)

Clear force interrupt. This bit is not cleared by hardware and must be manually cleared to stop the interrupt from continuing to be asserted.

impl<S: SliceId, M: SliceMode + ValidSliceMode<S>> Slice<S, M>

pub fn output_to<G: PinId + BankPinId + ValidPwmOutputPin<S, C>, PM: PinMode + ValidPinMode<G>, C: ChannelId>(
&mut self,
pin: Pin<G, PM>
) -> PwmPinToken<G>

Capture a gpio pin and use it as pwm output

impl<S: SliceId, M: SliceMode + ValidSliceInputMode<S>> Slice<S, M>

pub fn input_from<G: PinId + BankPinId + ValidPwmInputPin<S>, PM: PinMode + ValidPinMode<G>>(
&mut self,
pin: Pin<G, PM>
) -> PwmPinToken<G>

Capture a gpio pin and use it as pwm input for channel B