Struct rp2040_hal::rosc::RingOscillator

pub struct RingOscillator<S: State> { /* private fields */ }

A Ring Oscillator.

Implementations

impl<S: State> RingOscillator<S>

pub fn free(self) -> ROSC

Releases the underlying device.

impl RingOscillator<Disabled>

pub fn new(dev: ROSC) -> Self

Creates a new RingOscillator from the underlying device.

pub fn initialize(self) -> RingOscillator<Enabled>

Initializes the ROSC : frequency range is set, startup delay is calculated and set.

impl RingOscillator<Enabled>

pub fn operating_frequency(&self) -> HertzU32

Approx operating frequency of the ROSC in hertz

pub fn disable(self) -> RingOscillator<Disabled>

Disables the ROSC

pub fn get_random_bit(&self) -> bool

Generate random bit based on the Ring oscillator This is not suited for security purposes

pub unsafe fn dormant(self) -> RingOscillator<Dormant>

Put the ROSC in DORMANT state.

Safety

This method is marked unsafe because prior to switch the ROSC into DORMANT state, PLLs must be stopped and IRQs have to be properly configured. This method does not do any of that, it merely switches the ROSC to DORMANT state. See Chapter 2, Section 16, 5) for details.