Module rp2040_hal::sio

Single Cycle Input and Output (SIO)

To be able to partition parts of the SIO block to other modules:

use rp2040_hal::{gpio::Pins, pac, sio::Sio};

let mut peripherals = pac::Peripherals::take().unwrap();
let sio = Sio::new(peripherals.SIO);

And then for example

let pins = Pins::new(peripherals.IO_BANK0, peripherals.PADS_BANK0, sio.gpio_bank0, &mut peripherals.RESETS);

Structs

DivResult

Result of divide/modulo operation

HwDivider

Marker struct for ownership of divide/modulo module

Interp0

Interpolator Interp0

Interp0Lane0

The lane 0 of Interp0

Interp0Lane1

The lane 1 of Interp0

Interp1

Interpolator Interp1

Interp1Lane0

The lane 0 of Interp1

Interp1Lane1

The lane 1 of Interp1

LaneCtrl

Configuration struct for one lane of the interpolator

Sio

Struct containing ownership markers for managing ownership of the SIO registers.

SioFifo

Marker struct for ownership of SIO FIFO

SioGpioBank0

Marker struct for ownership of SIO gpio bank0

SioGpioQspi

Marker struct for ownership of SIO gpio qspi

Spinlock

Hardware based spinlock.

Traits

Interp

Trait representing the functionnality of an interpolator.

Lane

Trait representing the functionnality of a single lane of an interpolator.

SpinlockValid

This type is just used to limit us to Spinlocks 0..=31

Functions

spinlock_reset⚠

Free all spinlocks, regardless of their current status

spinlock_state

Returns the current state of the spinlocks. Each index corresponds to the associated spinlock, e.g. if index 5 is set to true, it means that Spinlock5 is currently locked.

Type Definitions

Spinlock0

Spinlock number 0

Spinlock1

Spinlock number 1

Spinlock2

Spinlock number 2

Spinlock3

Spinlock number 3

Spinlock4

Spinlock number 4

Spinlock5

Spinlock number 5

Spinlock6

Spinlock number 6

Spinlock7

Spinlock number 7

Spinlock8

Spinlock number 8

Spinlock9

Spinlock number 9

Spinlock10

Spinlock number 10

Spinlock11

Spinlock number 11

Spinlock12

Spinlock number 12

Spinlock13

Spinlock number 13

Spinlock14

Spinlock number 14

Spinlock15

Spinlock number 15

Spinlock16

Spinlock number 16

Spinlock17

Spinlock number 17

Spinlock18

Spinlock number 18

Spinlock19

Spinlock number 19

Spinlock20

Spinlock number 20

Spinlock21

Spinlock number 21

Spinlock22

Spinlock number 22

Spinlock23

Spinlock number 23

Spinlock24

Spinlock number 24

Spinlock25

Spinlock number 25

Spinlock26

Spinlock number 26

Spinlock27

Spinlock number 27

Spinlock28

Spinlock number 28

Spinlock29

Spinlock number 29

Spinlock30

Spinlock number 30