Trait rp2040_hal::sio::Lane

pub trait Lane {
    fn pop(&mut self) -> u32;
    fn peek(&self) -> u32;
    fn set_accum(&mut self, v: u32);
    fn get_accum(&self) -> u32;
    fn set_base(&mut self, v: u32);
    fn get_base(&self) -> u32;
    fn set_ctrl(&mut self, v: u32);
    fn get_ctrl(&self) -> u32;
    fn add_accum(&mut self, v: u32);
    fn read_raw(&self) -> u32;
}

Trait representing the functionnality of a single lane of an interpolator.

Required Methods

fn pop(&mut self) -> u32

Read the lane result, and simultaneously write lane results to both accumulators.

fn peek(&self) -> u32

Read the lane result without altering any internal state

fn set_accum(&mut self, v: u32)

Write a value to the accumulator

fn get_accum(&self) -> u32

Read the value from the accumulator

fn set_base(&mut self, v: u32)

Write a value to the base register

fn get_base(&self) -> u32

Read the value from the base register

fn set_ctrl(&mut self, v: u32)

Write to the control register

fn get_ctrl(&self) -> u32

Read from the control register

fn add_accum(&mut self, v: u32)

Add the value to the accumulator register

fn read_raw(&self) -> u32

Read the raw shift and mask value (BASE register not added)