Trait rp2040_hal::timer::Alarm

pub trait Alarm {
    fn clear_interrupt(&mut self);
    fn enable_interrupt(&mut self);
    fn disable_interrupt(&mut self);
    fn schedule(
        &mut self,
        countdown: MicrosDurationU32
    ) -> Result<(), ScheduleAlarmError>;
    fn schedule_at(
        &mut self,
        timestamp: Instant
    ) -> Result<(), ScheduleAlarmError>;
    fn finished(&self) -> bool;
}

Alarm abstraction.

Required Methods

fn clear_interrupt(&mut self)

Clear the interrupt flag.

The interrupt is unable to trigger a 2nd time until this interrupt is cleared.

fn enable_interrupt(&mut self)

Enable this alarm to trigger an interrupt.

After this interrupt is triggered, make sure to clear the interrupt with clear_interrupt.

fn disable_interrupt(&mut self)

Disable this alarm, preventing it from triggering an interrupt.

fn schedule(
&mut self,
countdown: MicrosDurationU32
) -> Result<(), ScheduleAlarmError>

Schedule the alarm to be finished after countdown. If enable_interrupt is called, this will trigger interrupt whenever this time elapses.

fn schedule_at(&mut self, timestamp: Instant) -> Result<(), ScheduleAlarmError>

Schedule the alarm to be finished at the given timestamp. If enable_interrupt is called, this will trigger interrupt whenever this timestamp is reached.

The RP2040 is unable to schedule an event taking place in more than u32::max_value() microseconds.

fn finished(&self) -> bool

Return true if this alarm is finished.