Struct rp2040_hal::uart::Pins

pub struct Pins<TX, RX, CTS, RTS> {
    pub tx: TX,
    pub rx: RX,
    pub rts: RTS,
    pub cts: CTS,
}

Customizable Uart pinout, allowing you to set the pins individually.

The following pins are valid UART pins:

UART	TX	RX	CTS	RTS

UART0	0, 12, 16, 28	1, 13, 17, 29	2, 14, 18	3, 15, 19

UART1	4, 8, 20, 24	5, 9, 21, 25	6, 10, 22, 26	7, 11, 23, 27

Every field can be set to () to not configure them.

Note that you can also use tuples (RX, TX) or (RX, TX, CTS, RTS) instead of this type.

This struct can either be filled manually or with a builder pattern:

let pins = Pins::default()
    .tx(gpio_pins.gpio0.into_mode())
    .rx(gpio_pins.gpio1.into_mode());

fn assert_is_valid_uart0<T: ValidUartPinout<UART0>>(_: T) {}

assert_is_valid_uart0(pins);

Fields
tx: TXrx: RXrts: RTScts: CTS
Implementations

impl<TX, RX, CTS, RTS> Pins<TX, RX, CTS, RTS>

pub fn tx<NTX>(self, tx: NTX) -> Pins<NTX, RX, CTS, RTS>

Set the TX pin

pub fn rx<NRX>(self, rx: NRX) -> Pins<TX, NRX, CTS, RTS>

Set the RX pin

pub fn cts<NCTS>(self, cts: NCTS) -> Pins<TX, RX, NCTS, RTS>

Set the CTS pin

pub fn rts<NRTS>(self, rts: NRTS) -> Pins<TX, RX, CTS, NRTS>

Set the RTS pin