Module rp2040_hal::xosc

Crystal Oscillator (XOSC)

Structs

CrystalOscillator

A Crystal Oscillator.

Disabled

XOSC is disabled (typestate)

Dormant

XOSC is in dormant mode (see Chapter 2, Section 16, 5)

Initialized

XOSC is initialized, ie we’ve given parameters (typestate)

Stable

Stable state (typestate)

StableOscillatorToken

A token that’s given when the oscillator is stablilzed, and can be exchanged to proceed to the next stage.

Enums

Error

Possible errors when initializing the CrystalOscillator

Traits

State

State of the Crystal Oscillator (typestate trait)

Functions

setup_xosc_blocking

Blocking helper method to setup the XOSC without going through all the steps.