Struct rp2040_hal::xosc::CrystalOscillator

pub struct CrystalOscillator<S: State> { /* private fields */ }

A Crystal Oscillator.

Implementations

impl<S: State> CrystalOscillator<S>

pub fn free(self) -> XOSC

Releases the underlying device.

impl CrystalOscillator<Disabled>

pub fn new(dev: XOSC) -> Self

Creates a new CrystalOscillator from the underlying device.

pub fn initialize(
self,
frequency: HertzU32
) -> Result<CrystalOscillator<Initialized>, Error>

Initializes the XOSC : frequency range is set, startup delay is calculated and set.

impl CrystalOscillator<Initialized>

pub fn await_stabilization(&self) -> Result<StableOscillatorToken, Infallible>

One has to wait for the startup delay before using the oscillator, ie awaiting stablilzation of the XOSC

pub fn get_stable(
self,
_token: StableOscillatorToken
) -> CrystalOscillator<Stable>

Returns the stablilzed oscillator

impl CrystalOscillator<Stable>

pub fn operating_frequency(&self) -> HertzU32

Operating frequency of the XOSC in hertz

pub fn disable(self) -> CrystalOscillator<Disabled>

Disables the XOSC

pub unsafe fn dormant(self) -> CrystalOscillator<Dormant>

Put the XOSC in DORMANT state.

Safety

This method is marked unsafe because prior to switch the XOSC into DORMANT state, PLLs must be stopped and IRQs have to be properly configured. This method does not do any of that, it merely switches the XOSC to DORMANT state. See Chapter 2, Section 16, 5) for details.